requirements.txt
# requirements.txt
google-generativeai>=0.7.0
pytest>=7.0.0  # Kept from original for future tests
requests>=2.28.0
python-dotenv>=1.0.0